+++ src/nn_laser_stabilizer/envs/neural_pid_env.py
@@ -25,7 +25,7 @@ class NeuralPIDEnv(BaseEnv):
         base_logger: Logger,
         control_min: int,
         control_max: int,
-        process_variable_max: float,
+        process_variable_max: int,
     ):
         super().__init__()
